pyproject.toml
[tool.poetry]
name = "qualitycontrolplatformtest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
